So what I didn't mention in the last commit was that joining was pulled out. It wasn't great anyhow, but now there's a proper joiner bod. She goes around looking for neighbouring cells which were dug by someone she doesn't know, so she then digs a door through, and ensures that the path is clear. It makes for better connected knotworks...
One issue with zoomorphs is that the knots are more broken anyway (every head/tail pair
is a dead-end). I may add an analysis function which goes through the knotwork ensuring
that there is only one loop - but it will still be broken by heads/tails.

Anyhow. This is much better..